test: Rename pdre-api to host-api adapter flag
