+++ test/adapters/substrate-legacy/src/pdre_api/child_storage.rs
@@ -1,4 +1,4 @@
-use crate::pdre_api::utils::{Decoder, ParsedInput, Runtime};
+use crate::host_api::utils::{Decoder, ParsedInput, Runtime};
 use parity_scale_codec::Encode;
 
 fn str<'a>(input: &'a [u8]) -> &'a str {
