+++ test/adapters/substrate-legacy/src/pdre_api/crypto.rs
@@ -1,4 +1,4 @@
-use crate::pdre_api::utils::{Decoder, ParsedInput, Runtime};
+use crate::host_api::utils::{Decoder, ParsedInput, Runtime};
 use parity_scale_codec::Encode;
 
 use sp_core::hashing::{twox_128, twox_256, twox_64};
