+++ test/adapters/substrate/src/host_api/crypto.rs
@@ -1,4 +1,4 @@
-use crate::pdre_api::utils::{str, Decoder, ParsedInput, Runtime};
+use crate::host_api::utils::{str, Decoder, ParsedInput, Runtime};
 use parity_scale_codec::Encode;
 use sp_core::crypto::key_types::DUMMY;
