+++ test/adapters/substrate/src/host_api/mod.rs
@@ -9,7 +9,7 @@ mod utils;
 use clap::ArgMatches;
 use utils::ParsedInput;
 
-pub fn process_pdre_api_tests(subcmd_matches: &ArgMatches) {
+pub fn process_host_api_tests(subcmd_matches: &ArgMatches) {
     if let Some(func) = subcmd_matches.value_of("function") {
         let input: ParsedInput = subcmd_matches.values_of("input").unwrap().into();
