+++ test/adapters/substrate/src/host_api/storage.rs
@@ -1,4 +1,4 @@
-use crate::pdre_api::utils::{str, Decoder, ParsedInput, Runtime};
+use crate::host_api::utils::{str, Decoder, ParsedInput, Runtime};
 use parity_scale_codec::Encode;
 
 pub fn ext_storage_set_version_1(input: ParsedInput) {
